refactor: remove compatibility layer
